Attempt to reconnect when waiting for activation
