+++ fuzz/attestation_fuzz.go
@@ -1,0 +1,1 @@
+package fuzz
