+++ validator/client/BUILD.bazel
@@ -16,6 +16,7 @@ go_library(
         "runner.go",
         "service.go",
         "validator.go",
+        "wait_for_activation.go",
     ],
     importpath = "github.com/prysmaticlabs/prysm/validator/client",
     visibility = ["//validator:__subpackages__"],
@@ -30,10 +31,12 @@ go_library(
         "//shared/featureconfig:go_default_library",
         "//shared/grpcutils:go_default_library",
         "//shared/hashutil:go_default_library",
+        "//shared/mathutil:go_default_library",
         "//shared/params:go_default_library",
         "//shared/rand:go_default_library",
         "//shared/slotutil:go_default_library",
         "//shared/timeutils:go_default_library",
+        "//shared/traceutil:go_default_library",
         "//validator/accounts/wallet:go_default_library",
         "//validator/db:go_default_library",
         "//validator/db/kv:go_default_library",
@@ -79,6 +82,7 @@ go_test(
         "runner_test.go",
         "service_test.go",
         "validator_test.go",
+        "wait_for_activation_test.go",
     ],
     embed = [":go_default_library"],
     deps = [
@@ -103,6 +107,7 @@ go_test(
         "@com_github_gogo_protobuf//types:go_default_library",
         "@com_github_golang_mock//gomock:go_default_library",
         "@com_github_hashicorp_golang_lru//:go_default_library",
+        "@com_github_pkg_errors//:go_default_library",
         "@com_github_prysmaticlabs_ethereumapis//eth/v1alpha1:go_default_library",
         "@com_github_prysmaticlabs_go_bitfield//:go_default_library",
         "@com_github_sirupsen_logrus//:go_default_library",
