+++ validator/client/validator_test.go
@@ -263,7 +263,6 @@ func TestWaitForChainStart_ReceiveErrorFromStream(t *testing.T) {
 	client := mock.NewMockBeaconNodeValidatorClient(ctrl)
 
 	v := validator{
-		//keyManager:      testKeyManager,
 		validatorClient: client,
 	}
 	clientStream := mock.NewMockBeaconNodeValidator_WaitForChainStartClient(ctrl)
@@ -280,172 +279,6 @@ func TestWaitForChainStart_ReceiveErrorFromStream(t *testing.T) {
 	assert.ErrorContains(t, want, err)
 }
 
-func TestWaitActivation_ContextCanceled(t *testing.T) {
-	ctrl := gomock.NewController(t)
-	defer ctrl.Finish()
-	client := mock.NewMockBeaconNodeValidatorClient(ctrl)
-	privKey, err := bls.RandKey()
-	require.NoError(t, err)
-	pubKey := [48]byte{}
-	copy(pubKey[:], privKey.PublicKey().Marshal())
-	km := &mockKeymanager{
-		keysMap: map[[48]byte]bls.SecretKey{
-			pubKey: privKey,
-		},
-	}
-	v := validator{
-		validatorClient: client,
-		keyManager:      km,
-	}
-	clientStream := mock.NewMockBeaconNodeValidator_WaitForActivationClient(ctrl)
-
-	client.EXPECT().WaitForActivation(
-		gomock.Any(),
-		&ethpb.ValidatorActivationRequest{
-			PublicKeys: [][]byte{pubKey[:]},
-		},
-	).Return(clientStream, nil)
-	clientStream.EXPECT().Recv().Return(
-		&ethpb.ValidatorActivationResponse{},
-		nil,
-	)
-	ctx, cancel := context.WithCancel(context.Background())
-	cancel()
-	assert.ErrorContains(t, cancelledCtx, v.WaitForActivation(ctx))
-}
-
-func TestWaitActivation_StreamSetupFails(t *testing.T) {
-	ctrl := gomock.NewController(t)
-	defer ctrl.Finish()
-	client := mock.NewMockBeaconNodeValidatorClient(ctrl)
-	privKey, err := bls.RandKey()
-	require.NoError(t, err)
-	pubKey := [48]byte{}
-	copy(pubKey[:], privKey.PublicKey().Marshal())
-	km := &mockKeymanager{
-		keysMap: map[[48]byte]bls.SecretKey{
-			pubKey: privKey,
-		},
-	}
-	v := validator{
-		validatorClient: client,
-		keyManager:      km,
-	}
-	clientStream := mock.NewMockBeaconNodeValidator_WaitForActivationClient(ctrl)
-	client.EXPECT().WaitForActivation(
-		gomock.Any(),
-		&ethpb.ValidatorActivationRequest{
-			PublicKeys: [][]byte{pubKey[:]},
-		},
-	).Return(clientStream, errors.New("failed stream"))
-	err = v.WaitForActivation(context.Background())
-	want := "could not setup validator WaitForActivation streaming client"
-	assert.ErrorContains(t, want, err)
-}
-
-func TestWaitActivation_ReceiveErrorFromStream(t *testing.T) {
-	ctrl := gomock.NewController(t)
-	defer ctrl.Finish()
-	client := mock.NewMockBeaconNodeValidatorClient(ctrl)
-
-	privKey, err := bls.RandKey()
-	require.NoError(t, err)
-	pubKey := [48]byte{}
-	copy(pubKey[:], privKey.PublicKey().Marshal())
-	km := &mockKeymanager{
-		keysMap: map[[48]byte]bls.SecretKey{
-			pubKey: privKey,
-		},
-	}
-	v := validator{
-		validatorClient: client,
-		keyManager:      km,
-	}
-	clientStream := mock.NewMockBeaconNodeValidator_WaitForActivationClient(ctrl)
-	client.EXPECT().WaitForActivation(
-		gomock.Any(),
-		&ethpb.ValidatorActivationRequest{
-			PublicKeys: [][]byte{pubKey[:]},
-		},
-	).Return(clientStream, nil)
-	clientStream.EXPECT().Recv().Return(
-		nil,
-		errors.New("fails"),
-	)
-	err = v.WaitForActivation(context.Background())
-	want := "could not receive validator activation from stream"
-	assert.ErrorContains(t, want, err)
-}
-
-func TestWaitActivation_LogsActivationEpochOK(t *testing.T) {
-	hook := logTest.NewGlobal()
-	ctrl := gomock.NewController(t)
-	defer ctrl.Finish()
-	client := mock.NewMockBeaconNodeValidatorClient(ctrl)
-	privKey, err := bls.RandKey()
-	require.NoError(t, err)
-	pubKey := [48]byte{}
-	copy(pubKey[:], privKey.PublicKey().Marshal())
-	km := &mockKeymanager{
-		keysMap: map[[48]byte]bls.SecretKey{
-			pubKey: privKey,
-		},
-	}
-	v := validator{
-		validatorClient: client,
-		keyManager:      km,
-		genesisTime:     1,
-	}
-	resp := generateMockStatusResponse([][]byte{pubKey[:]})
-	resp.Statuses[0].Status.Status = ethpb.ValidatorStatus_ACTIVE
-	clientStream := mock.NewMockBeaconNodeValidator_WaitForActivationClient(ctrl)
-	client.EXPECT().WaitForActivation(
-		gomock.Any(),
-		&ethpb.ValidatorActivationRequest{
-			PublicKeys: [][]byte{pubKey[:]},
-		},
-	).Return(clientStream, nil)
-	clientStream.EXPECT().Recv().Return(
-		resp,
-		nil,
-	)
-	assert.NoError(t, v.WaitForActivation(context.Background()), "Could not wait for activation")
-	require.LogsContain(t, hook, "Validator activated")
-}
-
-func TestWaitForActivation_Exiting(t *testing.T) {
-	ctrl := gomock.NewController(t)
-	defer ctrl.Finish()
-	client := mock.NewMockBeaconNodeValidatorClient(ctrl)
-	privKey, err := bls.RandKey()
-	require.NoError(t, err)
-	pubKey := [48]byte{}
-	copy(pubKey[:], privKey.PublicKey().Marshal())
-	km := &mockKeymanager{
-		keysMap: map[[48]byte]bls.SecretKey{
-			pubKey: privKey,
-		},
-	}
-	v := validator{
-		validatorClient: client,
-		keyManager:      km,
-		genesisTime:     1,
-	}
-	resp := generateMockStatusResponse([][]byte{pubKey[:]})
-	resp.Statuses[0].Status.Status = ethpb.ValidatorStatus_EXITING
-	clientStream := mock.NewMockBeaconNodeValidator_WaitForActivationClient(ctrl)
-	client.EXPECT().WaitForActivation(
-		gomock.Any(),
-		&ethpb.ValidatorActivationRequest{
-			PublicKeys: [][]byte{pubKey[:]},
-		},
-	).Return(clientStream, nil)
-	clientStream.EXPECT().Recv().Return(
-		resp,
-		nil,
-	)
-	require.NoError(t, v.WaitForActivation(context.Background()))
-}
 
 func TestCanonicalHeadSlot_FailedRPC(t *testing.T) {
 	ctrl := gomock.NewController(t)
